add remove offset

+++ scripts/train/train_ndvi_evidential.py
@@ -60,7 +60,7 @@ def main(args: Namespace):
         callbacks=default_callbacks(), 
         accelerator='gpu', 
         devices=1,
-        precision=32,
+        precision=16,
         max_epochs=args.max_epochs,
         default_root_dir=args.checkpoint_path,
         log_every_n_steps=300,
